Split collider and physics logic
