+++ Castaway.Level/Controllers/Colliders/ColliderBoxController.cs
@@ -1,4 +1,3 @@
-using System;
 using BepuPhysics;
 using BepuPhysics.Collidables;
 using Castaway.Math;
@@ -6,166 +5,24 @@ using Castaway.Math;
 namespace Castaway.Level.Controllers.Colliders;
 
 [ControllerName("Collider.Box")]
-public class ColliderBoxController : Controller
+public class ColliderBoxController : Controller, ICollider
 {
-	private Simulation? _simulation;
-
-	public BodyHandle Body;
 	[LevelSerialized("Mass")] public float Mass = 1f;
-	[LevelSerialized("Mode")] public PhysicsMode PhysicsMode = PhysicsMode.Dynamic;
-	public TypedIndex Shape;
 	[LevelSerialized("Size")] public Vector3 Size = new(1, 1, 1);
-	public StaticHandle Static;
 
-	public override Vector3 Position
-	{
-		set
-		{
-			if (_simulation == null) return;
-			switch (PhysicsMode)
-			{
-				case PhysicsMode.Dynamic:
-				{
-					_simulation.Bodies.GetDescription(Body, out var desc);
-					desc.Pose.Position =
-						new System.Numerics.Vector3((float)value.X, (float)value.Y, (float)value.Z);
-					_simulation.Bodies.ApplyDescription(Body, desc);
-					break;
-				}
-				case PhysicsMode.Static:
-				{
-					_simulation.Statics.GetDescription(Static, out var desc);
-					desc.Pose.Position =
-						new System.Numerics.Vector3((float)value.X, (float)value.Y, (float)value.Z);
-					_simulation.Statics.ApplyDescription(Static, desc);
-					break;
-				}
-				default:
-					throw new ArgumentOutOfRangeException();
-			}
-		}
-	}
-
-	public override Quaternion Rotation
-	{
-		set
-		{
-			if (_simulation == null) return;
-			switch (PhysicsMode)
-			{
-				case PhysicsMode.Dynamic:
-				{
-					_simulation.Bodies.GetDescription(Body, out var desc);
-					desc.Pose.Orientation = new System.Numerics.Quaternion(
-						(float)value.X, (float)value.Y, (float)value.Z, (float)value.W);
-					_simulation.Bodies.ApplyDescription(Body, desc);
-					break;
-				}
-				case PhysicsMode.Static:
-				{
-					_simulation.Statics.GetDescription(Static, out var desc);
-					desc.Pose.Orientation = new System.Numerics.Quaternion(
-						(float)value.X, (float)value.Y, (float)value.Z, (float)value.W);
-					_simulation.Statics.ApplyDescription(Static, desc);
-					break;
-				}
-				default:
-					throw new ArgumentOutOfRangeException();
-			}
-		}
-	}
+	private Box Box => new((float)Size.X, (float)Size.Y, (float)Size.Z);
+	public TypedIndex Shape { get; private set; }
+	public BodyInertia Inertia => Box.ComputeInertia(Mass);
 
 	public override void OnInit(LevelObject parent)
 	{
 		base.OnInit(parent);
-		var sim = parent.Level.PhysicsSimulation;
-		var shape = new Box((float)Size.X, (float)Size.Y, (float)Size.Z);
-		var inertia = shape.ComputeInertia(Mass);
-		Shape = sim.Shapes.Add(shape);
-		var pos = parent.Position;
-		switch (PhysicsMode)
-		{
-			case PhysicsMode.Dynamic:
-			{
-				var desc = BodyDescription.CreateDynamic(
-					new System.Numerics.Vector3((float)pos.X, (float)pos.Y, (float)pos.Z),
-					inertia,
-					new CollidableDescription(Shape, 0.1f),
-					new BodyActivityDescription(0.01f));
-				desc.Pose.Orientation = new System.Numerics.Quaternion(
-					(float)parent.Rotation.X,
-					(float)parent.Rotation.Y,
-					(float)parent.Rotation.Z,
-					(float)parent.Rotation.W);
-				Body = sim.Bodies.Add(desc);
-				break;
-			}
-			case PhysicsMode.Static:
-			{
-				var desc = new StaticDescription(
-					new System.Numerics.Vector3((float)pos.X, (float)pos.Y, (float)pos.Z),
-					new System.Numerics.Quaternion(
-						(float)parent.Rotation.X,
-						(float)parent.Rotation.Y,
-						(float)parent.Rotation.Z,
-						(float)parent.Rotation.W),
-					Shape,
-					ContinuousDetection.Passive);
-				Static = sim.Statics.Add(desc);
-				break;
-			}
-			default:
-				throw new ArgumentOutOfRangeException();
-		}
-
-		_simulation = sim;
+		Shape = parent.Level.PhysicsSimulation.Shapes.Add(Box);
 	}
 
 	public override void OnDestroy(LevelObject parent)
 	{
 		base.OnDestroy(parent);
-		switch (PhysicsMode)
-		{
-			case PhysicsMode.Dynamic:
-				_simulation?.Bodies?.Remove(Body);
-				break;
-			case PhysicsMode.Static:
-				_simulation?.Statics?.Remove(Static);
-				break;
-			default:
-				throw new ArgumentOutOfRangeException();
-		}
-
-		_simulation?.Shapes?.Remove(Shape);
-	}
-
-	public override void PostUpdate(LevelObject parent)
-	{
-		base.PostUpdate(parent);
-		System.Numerics.Vector3 p;
-		var sim = parent.Level.PhysicsSimulation;
-		switch (PhysicsMode)
-		{
-			case PhysicsMode.Dynamic:
-			{
-				sim.Bodies.GetDescription(Body, out var desc);
-				p = desc.Pose.Position;
-				parent.Position = new Vector3(p.X, p.Y, p.Z);
-				var r = desc.Pose.Orientation;
-				parent.Rotation = new Quaternion(r.W, r.X, r.Y, r.Z);
-				break;
-			}
-			case PhysicsMode.Static:
-			{
-				sim.Statics.GetDescription(Static, out var desc);
-				p = desc.Pose.Position;
-				parent.Position = new Vector3(p.X, p.Y, p.Z);
-				var r = desc.Pose.Orientation;
-				parent.Rotation = new Quaternion(r.W, r.X, r.Y, r.Z);
-				break;
-			}
-			default:
-				throw new ArgumentOutOfRangeException();
-		}
+		parent.Level.PhysicsSimulation.Shapes.Remove(Shape);
 	}
 }
