+++ Castaway.Level/Controllers/Colliders/ICollider.cs
@@ -1,0 +1,10 @@
+using BepuPhysics;
+using BepuPhysics.Collidables;
+
+namespace Castaway.Level.Controllers.Colliders;
+
+public interface ICollider
+{
+	TypedIndex Shape { get; }
+	BodyInertia Inertia { get; }
+}
